Issue #28 - DIP.transferObjectValue TargetLogicalData raceCondition.
DIP.transferObjectValue was returning a LogicalData that was assigned as the targetData of the TransferObjectRequest.

transferObjectValue may release the semaphore of the TOR before the AP thread assigns the targetData. Thus, the main thread may continue the execution and raise a null pointer exception when accessing the targetData of the TOR.