Add extension to shm file.
This commit is mostly usefull for tests between using either System-V
shared memory of plain file with mmap(MAP_SHARED). However the later
seems to show trouble when adding a heavy load on the filesystem. Hence,
for now, the System-V limitations are not too complicated to overcome as
it is simply a question of setting the system's limits.